% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bundles.R
\name{all_bundles}
\alias{all_bundles}
\title{Retrieve all available bundles}
\usage{
all_bundles(version = "latest")
}
\arguments{
\item{version}{The version of the OMOP bundles to use. Defaults to \code{"latest"}.}
}
\value{
A data.frame with the available bundles, containing their \code{id}, \code{concept_name}, \code{domain}
and \code{version}.
}
\description{
This is essentially a wrapper around \code{\link[omopbundles:available_bundles]{omopbundles::available_bundles()}}
to retrieve the available bundles.
}
\keyword{internal}
