% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preprocessing-summarise.R
\name{calculate_monthly_counts}
\alias{calculate_monthly_counts}
\title{Calculate monthly statistics for an OMOP concept}
\usage{
calculate_monthly_counts(omop_table, concept, date)
}
\arguments{
\item{omop_table}{A table from the OMOP CDM}

\item{concept}{The name of the concept column to calculate statistics for}

\item{date}{The name of the date column to calculate statistics for}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{concept_id}: The concept ID
\item \code{concept_name}: The concept name
\item \code{date_year}: The year of the date
\item \code{date_month}: The month of the date
\item \code{person_count}: The number of unique patients per concept for each month
\item \code{records_per_person}: The average number of records per person per concept for each month
}
}
\description{
Calculate monthly statistics for an OMOP concept
}
