% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preprocessing-summarise.R
\name{calculate_summary_stats}
\alias{calculate_summary_stats}
\title{Calculate summary statistics for an OMOP table}
\usage{
calculate_summary_stats(omop_table, concept_name)
}
\arguments{
\item{omop_table}{A table from the OMOP CDM}

\item{concept_name}{The name of the concept ID column}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{concept_id}: The concept ID
\item \code{summary_attribute}: The summary attribute (e.g. "mean", "sd", "frequency")
\item \code{value_as_number}: The value of the summary attribute
\item \code{value_as_concept_id}: In case of a categorical concept, the concept ID for each category
}
}
\description{
Calculates the mean snd standard deviation for numeric concepts and the
frequency for categorical concepts.
}
