% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bundles.R
\name{get_bundle_concepts}
\alias{get_bundle_concepts}
\title{Retrieve concepts for a given bundle}
\usage{
get_bundle_concepts(id, domain)
}
\arguments{
\item{id}{The id of the bundle to retrieve concepts for.}

\item{domain}{The domain of the bundle to retrieve concepts for.}
}
\value{
An \code{integer} vector with the concept IDs.
}
\description{
This is essentially a wrapper around \code{\link[omopbundles:concept_by_bundle]{omopbundles::concept_by_bundle()}}
to retrieve the concepts belonging to a specific bundle, defined by its \code{id} and \code{domain},
but returning only the concept IDs as a vector.
}
\keyword{internal}
