% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocessing_db.R
\name{process_summary_stats}
\alias{process_summary_stats}
\title{Generate the 'omopcat_summary_stats' table}
\usage{
process_summary_stats(cdm)
}
\arguments{
\item{cdm}{A \code{\link{CDMConnector}} object, e.g. from \code{\link[CDMConnector:cdm_from_con]{CDMConnector::cdm_from_con()}}}
}
\value{
A \code{data.frame} with the summary statistics
}
\description{
Generate the 'omopcat_summary_stats' table
}
