% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocessing_db.R
\name{query_concepts_table}
\alias{query_concepts_table}
\title{Function to produce the 'omopcat_concepts' table from a list of concept ids}
\usage{
query_concepts_table(cdm, concepts)
}
\arguments{
\item{cdm}{A \code{\link{CDMConnector}} object, e.g. from \code{\link[CDMConnector:cdm_from_con]{CDMConnector::cdm_from_con()}}}

\item{concepts}{A vector of concept IDs}
}
\value{
A \code{data.frame} with the concept table
}
\description{
Function to produce the 'omopcat_concepts' table from a list of concept ids
}
