% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocessing_db.R
\name{read_parquet_sorted}
\alias{read_parquet_sorted}
\title{Read a parquet table and sort the results}
\usage{
read_parquet_sorted(path, options = nanoparquet::parquet_options())
}
\arguments{
\item{path}{path to the parquet file to be read}

\item{options}{Nanoparquet options, see \code{\link[nanoparquet:parquet_options]{parquet_options()}}.}
}
\value{
A \code{data.frame} with the results sorted by all columns
}
\description{
Read a parquet table and sort the results
}
