% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preprocessing-db.R
\name{read_table_sorted}
\alias{read_table_sorted}
\title{Read a table from the database and sort the results}
\usage{
read_table_sorted(con, table, schema)
}
\arguments{
\item{con}{A \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}} object}

\item{table}{character, name of the table to write to}

\item{schema}{character, name of the schema to be used}
}
\value{
A \code{data.frame} with the results sorted by all columns
}
\description{
Read a table from the database and sort the results
}
