% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preprocessing.R
\name{write_table}
\alias{write_table}
\title{Write data to a table in the database}
\usage{
write_table(data, con, table, schema)
}
\arguments{
\item{data}{data.frame, data to be written to the table}

\item{con}{A \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}} object}

\item{table}{character, name of the table to write to}

\item{schema}{character, name of the schema to be used}
}
\value{
\code{TRUE}, invisibly, if the operation was successful
}
\description{
Write data to a table in the database
}
