% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preprocessing.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to duckdb database}
\usage{
connect_to_db(db_path, ..., .envir = parent.frame())
}
\arguments{
\item{db_path}{path to the duckdb database file}

\item{...}{unused}

\item{.envir}{passed on to \code{\link[withr:defer]{withr::defer()}}}
}
\value{
A \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}} object
}
\description{
Connect to duckdb database
}
